test(metrics): isolate global meter listeners
Global MeterListener publication state races across parallel tests and can hide observable instruments.

Closes #1949
